DevAccountDeployer trait no longer necessary
