+++ examples/src/ref_finance.rs
@@ -2,7 +2,6 @@ use std::{collections::HashMap, convert::TryInto};
 
 use near_units::{parse_gas, parse_near};
 use workspaces::network::Sandbox;
-use workspaces::prelude::*;
 use workspaces::{Account, AccountId, Contract, Network, Worker};
 use workspaces::{BlockHeight, DevNetwork};
