+++ examples/src/status_message.rs
@@ -1,5 +1,4 @@
 use serde_json::json;
-use workspaces::prelude::*;
 
 const STATUS_MSG_WASM_FILEPATH: &str = "./examples/res/status_message.wasm";
