+++ workspaces/src/prelude.rs
@@ -1,3 +1,3 @@
 //! All traits that are essential to the ease of use of workspaces.
 
-pub use crate::network::DevAccountDeployer;
+pub use crate::network::TopLevelAccountCreator;
