+++ workspaces/src/worker/mod.rs
@@ -3,17 +3,15 @@ mod impls;
 use std::fmt;
 use std::sync::Arc;
 
-use crate::network::{Betanet, Mainnet, Sandbox, Testnet, TopLevelAccountCreator};
-use crate::result::CallExecution;
-use crate::types::{AccountId, SecretKey};
-use crate::{Account, Contract, DevNetwork, Network, Result};
+use crate::network::{Betanet, Mainnet, Sandbox, Testnet};
+use crate::{Network, Result};
 
 /// The `Worker` type allows us to interact with any NEAR related networks, such
 /// as mainnet and testnet. This controls where the environment the worker is
 /// running on top of is. Refer to this for all network related actions such as
 /// deploying a contract, or interacting with transactions.
 pub struct Worker<T: ?Sized> {
-    workspace: Arc<T>,
+    pub(crate) workspace: Arc<T>,
 }
 
 impl<T> Worker<T>
@@ -32,28 +30,6 @@ impl<T: Network + 'static> Worker<T> {
         Worker {
             workspace: self.workspace,
         }
-    }
-}
-
-impl<T> Worker<T>
-where
-    T: DevNetwork + TopLevelAccountCreator + 'static,
-{
-    pub async fn create_tla(&self, id: AccountId, sk: SecretKey) -> Result<CallExecution<Account>> {
-        self.workspace
-            .create_tla(self.clone().coerce(), id, sk)
-            .await
-    }
-
-    pub async fn create_tla_and_deploy(
-        &self,
-        id: AccountId,
-        sk: SecretKey,
-        wasm: &[u8],
-    ) -> Result<CallExecution<Contract>> {
-        self.workspace
-            .create_tla_and_deploy(self.clone().coerce(), id, sk, wasm)
-            .await
     }
 }
