+++ workspaces/tests/batch_tx.rs
@@ -1,7 +1,6 @@
 use serde_json::json;
 use test_log::test;
 use workspaces::operations::Function;
-use workspaces::prelude::*;
 
 #[test(tokio::test)]
 async fn test_batch_tx() -> anyhow::Result<()> {
