+++ workspaces/tests/create_account.rs
@@ -1,6 +1,5 @@
 #![recursion_limit = "256"]
 use test_log::test;
-use workspaces::prelude::*;
 
 #[test(tokio::test)]
 async fn test_subaccount_creation() -> anyhow::Result<()> {
